// src/Models/GamePieceType.ts
export enum GamePieceType {
  None = "None",
  KingVictor = "KingVictor",
  KingChallenger = "KingChallenger",
  Rook = "Rook",
  RookPro = "RookPro",
  Bishop = "Bishop",
  BishopPro = "BishopPro",
  Gold = "Gold",
  Silver = "Silver",
  SilverPro = "SilverPro",
  Knight = "Knight",
  KnightPro = "KnightPro",
  Lance = "Lance",
  LancePro = "LancePro",
  Pawn = "Pawn",
  PawnPro = "PawnPro"

}